Move core time parsing capabilities to coreinternal
